feat: add enriched error interface to flux
We need a common way of categorizing and understanding errors returned
from InfluxDB. We model the errors based on the same concepts that
InfluxDB uses, but we create our own error codes that are modeled on the
grpc error codes.

Each native flux error will have a code that represents the category or
type of error and a message that gives detailed and human-readable
output about the error.